Differentiate between results and no results
